% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.bcmc.R
\name{comp.bcmc}
\alias{comp.bcmc}
\title{The complete procedure of running BCMC (Biomarker Categorization in Meta-analysis
by Concordance) for biomarker detection and categorization.}
\usage{
comp.bcmc(
  data.exp,
  data.clin,
  meta.de = TRUE,
  de.lfc = NULL,
  de.pval = NULL,
  B = 100,
  parallel = FALSE,
  num.cores = NULL
)
}
\arguments{
\item{data.exp:}{The gene expression data. It should be a list in which
every element represents a study. Within each element (study), the data is
supposed to be matrix where each row is a gene and each column is a sample.}

\item{data.clin:}{The clinical data about whether a sample is case or control.
It should be a list in which every element represents a study. Within each
element (study), the data is supposed to be a data frame with only one column
and each row is a sample (same order as in \code{data.exp}). In each data
frame, code control as 0 and case a 1.}

\item{meta.de:}{A logical value indicating whether the \code{MetaDE} should
be run before the BCMC. The default is TRUE. If FALSE, users need to provide
the log fold change and p-value from DE analysis as arguments \code{de.lfc}
and \code{de.pval}.}

\item{de.lfc:}{If \code{meta.de} is FALSE, this should be a matrix of log fold
change from DE analysis. Each row should be a gene and each column should be
a study. Otherwise, the deafult is TRUE and \code{MetaDE} will be run for
DE analysis before the BCMC.}

\item{de.pval:}{If \code{meta.de} is FALSE, this should be a matrix of p-value
from DE analysis. Each row should be a gene and each column should be
a study. Otherwise, the deafult is TRUE and \code{MetaDE} will be run for
DE analysis before the BCMC.}

\item{B:}{Number of permutations. The default is 100.}

\item{parallel:}{A logical value indicating whether the parallel computing
should be done. The default is FALSE.}

\item{num.cores:}{If \code{parallel} is TRUE, this is the number of cores
to be used in the paralle computing.}
}
\value{
A list with following components:
\item{lfc}{The matrix of log fold change from DE analysis. Each row is a gene
and each column is a study.}
\item{pval}{The matrix of p-value from DE analysis. Each row is a gene
and each column is a study.}
\item{pos.wp}{The matrix of predicted up-regulated weight pattern. Each row
is a gene and each column is a study.}
\item{neg.wp}{The matrix of predicted down-regulated weight pattern. Each row
is a gene and each column is a study.}
\item{max.wp}{The matrix of predicted dominant weight pattern. Each row
is a gene and each column is a study. This is found by comparing the two BCMC
statistics for up-regualted and down-regulated.}
\item{Rg}{A matrix of BCMC statistcs, as the \eqn{R_g} in the paper. Each row
is a gene. The first column is for up-regulated pattern, the second column
is for down-regulated pattern, and the third column is the larger one of the
previous two statistics. See the paper for more details.}
\item{pvalue}{A matrix of permutation p-values where each row is a gene. The
first column is for up-regulated statistc, the second column is for down
-regulated statistic, and the third colum is for the dominant statistic.}
\item{qvalue}{A matrix of permutation q-values where each row is a gene. The
first column is for up-regulated statistc, the second column is for down
-regulated statistic, and the third colum is for the dominant statistic.}
}
\description{
This function will first run the \code{MetaDE} to perform DE analysis, then
implement the meta-analysis method BCMC for biomarker detection and categorization
by \code{bcmc} function, and calculate the permutation-based p-value and q-value
using \code{perm.bcmc} function.
}
\examples{
data("SimulDE")
result_comp <- comp.bcmc(data.exp=SimulDE$express, data.clin=SimulDE$clin, 
                         B=5, parallel=FALSE)
names(result_comp)
head(result_comp$Rg)
head(result_comp$pos.wp)
head(result_comp$pvalue)
}
