% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.bcmc.R
\name{perm.bcmc}
\alias{perm.bcmc}
\title{Calculate the permutation-based p-value and FDR adjusted p-value (also
known as q-value) after BCMC.}
\usage{
perm.bcmc(
  B = 100,
  data.exp,
  data.clin,
  res.bcmc,
  parallel = FALSE,
  num.cores = NULL
)
}
\arguments{
\item{B:}{Number of permutations. The default is 100.}

\item{data.exp:}{The gene expression data. It should be a list in which
every element represents a study. Within each element (study), the data is
supposed to be matrix where each row is a gene and each column is a sample.}

\item{data.clin:}{The clinical data about whether a sample is case or control.
It should be a list in which every element represents a study. Within each
element (study), the data is supposed to be a data frame with only one column
and each row is a sample (same order as in \code{data.exp}). In each data
frame, code control as 0 and case a 1.}

\item{res.bcmc:}{The results of running BCMC. Typically the output of
\code{bcmc} function.}

\item{parallel:}{A logical value indicating whether the parallel computing
should be done. The default is FALSE.}

\item{num.cores:}{If \code{parallel} is TRUE, this is the number of cores
to be used in the paralle computing.}
}
\value{
A list with following components:
\item{pvalue}{A matrix of permutation p-values where each row is a gene. The
first column is for up-regulated statistc, the second column is for down
-regulated statistic, and the third colum is for the dominant statistic.}
\item{qvalue}{A matrix of permutation q-values where each row is a gene. The
first column is for up-regulated statistc, the second column is for down
-regulated statistic, and the third colum is for the dominant statistic.}
}
\description{
This function use a permutation-based test to calculate the p-value and
FDR adjusted p-value (also known as q-value) after doing the BCMC.
Details can be found in the paper.
}
